Restructure project hierarchy by function
