+++ src/modules/game/components/GameFieldComponent.tsx
@@ -1,15 +1,15 @@
 import * as React from "react";
 
-import Cell from "./cell";
+import GameCellComponent from "./GameCellComponent";
 
-interface IBoardProps {
+interface GameFieldProps {
   numbers: number[][];
   onCellClick: (row: number, col: number) => void;
   previousSelectedNumberRow?: number;
   previousSelectedNumberCol?: number;
 }
 
-class Board extends React.Component<IBoardProps> {
+class GameFieldComponent extends React.Component<GameFieldProps> {
   get numbers() {
     return this.props.numbers;
   }
@@ -26,10 +26,10 @@ class Board extends React.Component<IBoardProps> {
                 const crossedOut = !selected && theNumber === 0;
                 return (
                   <div className="col-" key={(rowIndex << 10) + colIndex}>
-                    <Cell onClick={() => this.props.onCellClick(rowIndex, colIndex)}
-                          selected={selected}
-                          crossedOut={crossedOut}
-                          value={theNumber.toString()}
+                    <GameCellComponent onClick={() => this.props.onCellClick(rowIndex, colIndex)}
+                                       selected={selected}
+                                       crossedOut={crossedOut}
+                                       value={theNumber.toString()}
                     />
                   </div>
                 );
@@ -42,4 +42,4 @@ class Board extends React.Component<IBoardProps> {
   }
 }
 
-export default Board;
+export default GameFieldComponent;
