+++ tslint.json
@@ -10,7 +10,8 @@
     "member-ordering": false,
     "trailing-comma": false,
     "object-literal-sort-keys": false,
-    "semicolon": false
+    "semicolon": false,
+    "interface-name": false
   },
   "jsRules": {
     "quotemark": false
